+++ lib/whois/record/parser/whois.rrpproxy.net.rb
@@ -22,22 +22,24 @@ module Whois
       #
       # @author Simone Carletti
       # @author Igor Dolzhikov <bluesriverz@gmail.com>
-      # @since  3.2.1
       #
       class WhoisRrpproxyNet < Base
         include Scanners::Scannable
 
         self.scanner = Scanners::BaseShared3
 
+
         property_supported :disclaimer do
           node("field:disclaimer")
         end
+
 
         property_supported :domain do
           node('DOMAIN', &:downcase)
         end
 
         property_not_supported :domain_id
+
 
         property_supported :status do
           if available?
@@ -56,10 +58,11 @@ module Whois
         end
 
         property_supported :registrar do
+          return unless registered?
           Record::Registrar.new(
-              :name => 'Key-Systems',
-              :organization => 'Key-Systems GmbH',
-              :url  => 'http://www.domaindiscount24.com/'
+              name:         'Key-Systems',
+              organization: 'Key-Systems GmbH',
+              url:          'http://www.domaindiscount24.com/'
           )
         end
 
@@ -78,7 +81,7 @@ module Whois
         property_supported :nameservers do
           node('nameserver') do |array|
             array.map do |name|
-              Nameserver.new(:name => name)
+              Nameserver.new(name: name)
             end
           end
         end
@@ -88,18 +91,18 @@ module Whois
         def build_contact(element, type)
           node("#{element}-contact") do
             Record::Contact.new(
-                :type         => type,
-                :id           => node("#{element}-contact"),
-                :name         => [node("#{element}-fname"), node("#{element}-lname")].join(' '),
-                :organization => node("#{element}-organization"),
-                :address      => node("#{element}-street"),
-                :city         => node("#{element}-city"),
-                :zip          => node("#{element}-zip"),
-                :state        => nil,
-                :country_code => node("#{element}-country"),
-                :phone        => node("#{element}-phone"),
-                :fax          => node("#{element}-fax"),
-                :email        => node("#{element}-email")
+                type:         type,
+                id:           node("#{element}-contact"),
+                name:         [node("#{element}-fname"), node("#{element}-lname")].join(' '),
+                organization: node("#{element}-organization"),
+                address:      node("#{element}-street"),
+                city:         node("#{element}-city"),
+                zip:          node("#{element}-zip"),
+                state:        nil,
+                country_code: node("#{element}-country"),
+                phone:        node("#{element}-phone"),
+                fax:          node("#{element}-fax"),
+                email:        node("#{element}-email")
             )
           end
         end